+++ docs/api/changelog.rst
@@ -22,6 +22,9 @@ Version 0.10
   is deprecated and you should upgrade your code to use one of the two named
   classes explicitly.
 
+* Make :meth:`aioxmpp.muc.Room.send_message_tracked` a normal method instead
+  of a coroutine (it was never intended to be a coroutine).
+
 .. _api-changelog-0.9:
 
 Version 0.9